app/requirements.txt
aerich==0.2.2
async-generator==1.10
async-timeout==3.0.1
asyncclick==7.0.9
asyncpg==0.21.0
bcrypt==3.1.7
email-validator==1.1.1
emails==0.6
fastapi
pydantic==1.6.2
PyJWT==2.4.0
starlette
tortoise-orm==0.16.14
typing-extensions==3.7.4.2
uvicorn==0.11.7
